requirements.txt
appdirs==1.4.4
black==21.7b0
cfgv==3.3.1
click==8.1.3
colorama==0.4.5
distlib==0.3.6
dnspython==2.2.1
filelock==3.8.0
identify==2.5.5
isort==5.5.2
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.23.3
pandas==1.4.4
pathspec==0.10.1
platformdirs==2.5.2
pre-commit==2.15.0
pydantic==1.9.0
pymongo==4.2.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2022.2.1
PyYAML==6.0
regex==2022.9.13
six==1.16.0
toml==0.10.2
tomli==1.2.3
typing-extensions==4.3.0
virtualenv==20.16.5
